Calculate Rank-based Triage Effectiveness (RTE)

Calculates Rank-based Triage Effectiveness (RTE) metrics, which evaluate triage performance based on patient queue positions rather than waiting times. RTE provides a patient-level metric that complements the waiting time-based Triage Effectiveness (WTE) approach.

Usage

calc_rte(
  df,
  subgroup = NULL,
  var1 = NULL,
  var2 = NULL,
  bootstrap = FALSE,
  bootstrap_params = list(sample_percentage = 1, n_iterations = 2000, distribution_span =
    0.95),
  n_workers = parallel::detectCores() - 1,
  verbose = FALSE,
  overall_only = FALSE,
  min_loset_warning = 5,
  include_distributions = TRUE,
  check_convergence = TRUE,
  seed = NULL,
  quiet = TRUE
)

Arguments

df

Data frame containing patient data (must be initialized through init())

subgroup

Optional list for subgroup analysis - an alternative to var1/2

var1

Optional string for first comparison variable

var2

Optional string for second comparison variable

bootstrap

Logical. Whether to perform bootstrap calculations. Default is FALSE.

bootstrap_params

List of bootstrap parameters:

sample_percentage (default: 1) How large of the sample is used? 1 = full sample.

n_iterations (default: 2000) How many bootstrap iterations to use

distribution_span (default: 0.95) What is the confidence interval used?

n_workers

Number of workers for parallel processing (default: detectCores() - 1)

verbose

Logical. If TRUE, returns patient-level RTE values for further analysis including intermediate calculation fields from calculate_queue_metrics. Default is FALSE.

overall_only

Logical. If TRUE, only overall metrics are returned (default: FALSE)

min_loset_warning

Numerical, when should the function warn for low loset prevalence? (default: 5)

include_distributions

Logical. If TRUE, returns bootstrap distributions. Default is FALSE.

check_convergence

Logical, Default is TRUE. Will check convergence and produce convergence plots if true + bootstrap == TRUE.

seed

Seed for reproducible bootstrapping (default: NULL - no seed)

quiet

Logical. Default is TRUE. If false outputs the recommended/used confidence interval choice.

Value

A list containing:

results: Tibble with RTE metrics

metadata: List containing calculation parameters

verbose_df: (If verbose) Patient-level RTE values

bootstrap_distributions: (If bootstrap & include_distribution) bootstrap distributions from bootstrapping.