Create Segments from Patient Data

This function divides the input patient data into segments based on continuous queue periods. Each segment represents a period where the queue starts at 0 and ends at 0, allowing for independent processing of these time periods.

Usage

create_segments(df, n_workers = detectCores() - 1)

Arguments

df

A data frame containing patient data, including ID, unit, arrival times, and resolve times.

n_workers

Integer. Number of cores to use for parallel processing. Default is (detectCores() - 1).

Value

A list of data frames, where each data frame represents a segment of continuous queue activity.

Details

The function identifies periods where the queue size goes from 0 to 1, marking the start of a new segment. It then groups patient data into these segments, which can be processed independently in parallel simulations. This segmentation is crucial for maintaining chronological integrity while enabling parallelization.