Create a Triage Effectiveness Plot

This function generates a plot to visualize Triage Effectiveness (TE) metrics. It can display Theoretical Triage Effectiveness (TTE) and Observed Triage Effectiveness (OTE) for different units, with optional grouping variables and variability intervals.

Usage

plot_te(
  data,
  title = "Triage Effectiveness Plot",
  show_tte = NULL,
  show_ote = NULL,
  show_var = NULL,
  var_alpha = 0.5,
  dumbell_width = 1,
  var_width = 1.5,
  var_end_width = 1,
  var_end_height = 0.1,
  label_style = "half",
  min_x = NULL
)

Arguments

data

A data frame containing the following columns:

unit: Factor or character vector of unit names

tte_te: Numeric vector of Theoretical Triage Effectiveness values

ote_te: Numeric vector of Observed Triage Effectiveness values

boot_tte_var_lower: Optional. Lower variability value for TTE

boot_tte_var_upper: Optional. Upper variability value for TTE

boot_ote_var_lower: Optional. Lower variability value for OTE

boot_ote_var_upper: Optional. Upper variability value for OTE

title

Character string for the plot title. Default is "Triage Effectiveness Plot".

show_tte

Logical, whether to display TTE values. Default is NULL. Will show if there is TTE data.

show_ote

Logical, whether to display OTE values. Default is NULL. Will show if there is OTE data.

show_var

Logical, whether to display variability intervals. Default is null. Will show if there is variability data supplied.

var_alpha

Numeric between 0 and 1, specifying the transparency of variability interval lines. Default is 0.5.

dumbell_width

Numeric, specifying the width of the line connecting TTE and OTE points. Default is 1. Note that since the line is horizontal, the "width" represents its thickness.

var_width

Numeric, specifying the width of the main variability interval lines - that is the vertical line at the end of the variability span. Default is 1.5.

var_end_width

Numeric, specifying the width (or thickness) of the variability interval end lines. Default is 1.

var_end_height

Numeric, specifying the height of variability interval end lines as offset from center. Default is 0.05.

label_style

categorical, valid inputs - "none", "small", "full" & "half". Decides the style of the label for TTE/OTE on the rights side of the graph. Default is half witch nets the label style as: "Theoretical TE". Small: "TTE", Full: "Theoretical Triage Effectiveness".

min_x

Percentual value that is the minimum value that the plot will stretch to. Default is null which uses the min value in the data rounded to the next 10%. Useful to set if you have several plots with one that has negative values to keep the same aspect ratio.

Value

A ggplot object representing the Triage Effectiveness plot.

Examples

if (FALSE) { # \dontrun{
data <- data.frame(
  unit = c("Unit A", "Unit B", "Unit C"),
  tte_te = c(0.6, 0.7, 0.5),
  ote_te = c(0.65, 0.75, 0.55),
  boot_tte_var_lower = c(0.55, 0.65, 0.45),
  boot_tte_var_upper = c(0.65, 0.75, 0.55),
  boot_ote_var_lower = c(0.60, 0.70, 0.50),
  boot_ote_var_upper = c(0.70, 0.80, 0.60)
)
plot_te(data, show_var = TRUE)
} # }